POST 기능 담당 객체추가, 기능이동 : MyNetworkWorker
